Fix for duplicate yellow letters
Sometimes, a user inputted word contains several time the same letter
But the word to guess contains this letter only once.
And in rare cases, none of the user word letter matches the correct position
in the word to guess.

Ex: to guess: PEGRE | user word: TARIR
None of the R will be green. But only one of them needs to be yellow.
The second one should be grey.